Basic face recognition added, testing the functionality
So far testing out the ability of the software to recognize a face. Once a face has been recognized, display it, if there is no face in the screen, display black colored image to cover everything. (What will later be a fail-safe for when the software doesn't detect face and doesn't blur it, in case the face moves in a weird way and gets exposed)